+++ clowder/cmd.py
@@ -11,12 +11,12 @@ import argcomplete
 import colorama
 from termcolor import cprint, colored
 
-import clowder.utility.formatting as fmt
+import clowder.util.formatting as fmt
 from clowder.clowder_controller import ClowderController
 from clowder.clowder_repo import ClowderRepo
 from clowder.error.clowder_error import ClowderError
-from clowder.utility.connectivity import is_offline
-from clowder.utility.subparsers import configure_argparse
+from clowder.util.connectivity import is_offline
+from clowder.util.subparsers import configure_argparse
 
 
 def main():
