+++ clowder/util/connectivity.py
