+++ clowder/util/formatting.py
